requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
python-multipart==0.0.6
openai==1.3.0
python-dateutil==2.8.2
regex==2023.10.3
langchain
google-genai